Add Global, Public and Private target configuration
Gesture recognizers in public targets are also combined with global targets.
Gesture recognizers in private targets are strictly local.
They are not merged with global ones.
Any target that is not explicitly defined public or private is considered to be public,
thus contains the whole set of globally defined gesture recognizers.